rag-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sentence-transformers>=3.1.1
supabase>=2.7.4,<2.10.0
groq>=0.11.0
python-dotenv>=1.0.1
httpx>=0.27.2
pydantic>=2.10.0
